员工排班设置bug fixed

+++ lichkin-projects-attendance-entities/src/main/java/com/lichkin/springframework/entities/impl/SysCompAttendanceAreaConfigEntity.java
@@ -34,7 +34,7 @@ public class SysCompAttendanceAreaConfigEntity extends CompIDEntity {
 
 	/** 水平面半径（50米） */
 	@Column(nullable = false)
-	private Byte radius;
+	private Integer radius;
 
 	/** 垂直面落差（米） */
 	@Column
